+++ NEWS.md
@@ -1,3 +1,7 @@
+# greenAlgoR 0.1.1
+
+* Remove the useless argument `fields` from `ga_targets()`
+
 # greenAlgoR 0.1
 
 * Initial github submission.